Rename "spatial" directory to "spatialAttachers"
