+++ packages/dev/core/src/AudioV2/abstractAudio/subNodes/spatialAudioSubNode.ts
@@ -3,8 +3,8 @@ import type { TransformNode } from "core/Meshes/transformNode";
 import type { Quaternion, Vector3 } from "../../../Maths/math.vector";
 import type { Nullable } from "../../../types";
 import type { AudioEngineV2 } from "../audioEngineV2";
-import { _SpatialAudioAttacher } from "../spatial/spatialAudioAttacher";
-import type { SpatialAudioAttachmentType } from "../spatial/spatialAudioAttacher";
+import { _SpatialAudioAttacher } from "../spatialAttachers/spatialAudioAttacher";
+import type { SpatialAudioAttachmentType } from "../spatialAttachers/spatialAudioAttacher";
 import type { ISpatialAudioOptions } from "../subProperties/abstractSpatialAudio";
 import { _SpatialAudioDefaults } from "../subProperties/abstractSpatialAudio";
 import type { _AbstractAudioSubGraph } from "./abstractAudioSubGraph";
