+++ packages/dev/core/src/AudioV2/abstractAudio/subProperties/spatialAudio.ts
@@ -1,7 +1,7 @@
 import type { Quaternion, Vector3 } from "../../../Maths/math.vector";
 import type { AbstractMesh, TransformNode } from "../../../Meshes";
 import type { Nullable } from "../../../types";
-import type { SpatialAudioAttachmentType } from "../spatial/spatialAudioAttacher";
+import type { SpatialAudioAttachmentType } from "../spatialAttachers/spatialAudioAttacher";
 import type { _AbstractAudioSubGraph } from "../subNodes/abstractAudioSubGraph";
 import { AudioSubNode } from "../subNodes/audioSubNode";
 import type { _SpatialAudioSubNode } from "../subNodes/spatialAudioSubNode";
